pyproject.toml
[project]
name = "grading-tool"
version = "0.1.0"
description = "Hybrid deterministic + Gemini grading tool for homework grading"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.33.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "pandas>=2.0.0",
  "google-generativeai>=0.8.0",
  "pyyaml>=6.0.0",
  "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"